Moved validate_path_parameter to Channel class
