+++ infobip_channels/core/channel.py
@@ -12,6 +12,7 @@ from infobip_channels.core.models import (
     PostHeaders,
     ResponseBase,
 )
+from infobip_channels.web_rtc.models.path_parameters.core import PathParameter
 
 
 class Channel(ABC):
@@ -113,6 +114,24 @@ class Channel(ABC):
         :return: Dictionary of the converted model
         """
         return model.dict(by_alias=by_alias, exclude_unset=exclude_unset, **kwargs)
+
+    @staticmethod
+    def validate_path_parameter(
+        parameter: Union[PathParameter, Dict], parameter_type: Type[PathParameter]
+    ) -> PathParameter:
+        """
+        Validate path parameter by trying to instantiate the provided class and
+        extract valid path parameter.
+
+        :param parameter: Path parameter to validate
+        :param parameter_type: Type of path parameter
+        :return: Class instance corresponding to the provided parameter type
+        """
+        return (
+            parameter
+            if isinstance(parameter, parameter_type)
+            else parameter_type(**parameter)
+        )
 
     @staticmethod
     def validate_message_body(
